+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.5] - 2026-02-24
+
+### Added
+- Added `getOrFetch` method for simple cache-or-fetch operations
+- Added `forceCache` parameter to `get` method
+
 ## [1.0.4] - 2026-01-07
 
 ### Added